combined outfit and clothing contexts into wardrobe context
